# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ArP - Artwork Provenance API
  description: |
    REST API for managing artwork provenance data using Semantic Web technologies.
    
    This API provides endpoints for:
    - Managing artworks (CRUD operations)
    - Tracking provenance chains (ownership history)
    - Executing SPARQL queries against the RDF knowledge base
    - Full-text search with faceted filtering
    - Artwork recommendations
    - QR code generation for artwork identification
    - **External SPARQL Integration**: Query DBpedia, Wikidata, and Getty AAT for enriched artwork metadata
  version: 1.0.0
  contact:
    name: ArP Development Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.arp.example.org/api
    description: Production server
...
paths:
...
  /artworks:
    get:
      tags: [artworks]
      summary: List all artworks
      description: Retrieve a paginated list of artworks with optional filtering and sorting.
      operationId: listArtworks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: artist
          in: query
          description: Filter by artist name
          schema:
            type: string
        - name: period
          in: query
          description: Filter by art historical period
          schema:
            type: string
        - name: search
          in: query
          description: Search term for quick filtering
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [title, dateCreated, -title, -dateCreated]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtworkListResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [artworks]
      summary: Create a new artwork
      description: Create a new artwork entry in the knowledge base.
      operationId: createArtwork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtworkCreate'
      responses:
        '201':
          description: Artwork created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artwork'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /artworks/{id}:
    parameters:
      - $ref: '#/components/parameters/ArtworkId'

    get:
      tags: [artworks]
      summary: Get artwork by ID
      description: Retrieve a single artwork by its unique identifier.
      operationId: getArtwork
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artwork'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [artworks]
      summary: Update artwork
      description: Update an existing artwork's information.
      operationId: updateArtwork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtworkUpdate'
      responses:
        '200':
          description: Artwork updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artwork'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [artworks]
      summary: Delete artwork
      description: Delete an artwork from the knowledge base.
      operationId: deleteArtwork
      responses:
        '204':
          description: Artwork deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # PROVENANCE CHAIN
  # ============================================
  /artworks/{id}/provenance:
    parameters:
      - $ref: '#/components/parameters/ArtworkId'

    get:
      tags: [provenance]
      summary: Get provenance chain
      description: Retrieve the complete provenance (ownership history) chain for an artwork.
      operationId: getProvenance
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProvenanceRecord'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [provenance]
      summary: Add provenance event
      description: Add a new provenance event to an artwork's history.
      operationId: addProvenanceEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvenanceEventCreate'
      responses:
        '201':
          description: Provenance event added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /artworks/{id}/provenance/{eventId}:
    parameters:
      - $ref: '#/components/parameters/ArtworkId'
      - $ref: '#/components/parameters/EventId'

    put:
      tags: [provenance]
      summary: Update provenance event
      description: Update an existing provenance event.
      operationId: updateProvenanceEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvenanceEventUpdate'
      responses:
        '200':
          description: Provenance event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [provenance]
      summary: Delete provenance event
      description: Remove a provenance event from an artwork's history.
      operationId: deleteProvenanceEvent
      responses:
        '204':
          description: Provenance event deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================
  # SPARQL QUERY
  # ============================================
...
  /artworks/{id}/enrich:
    parameters:
      - $ref: '#/components/parameters/ArtworkId'

    get:
      tags: [external]
      summary: Enrich artwork with external data
      description: |
        Fetch and combine external data for an artwork from DBpedia, Wikidata, and Getty AAT.
        
        This endpoint:
        1. Queries the local Fuseki triplestore for owl:sameAs links
        2. Fetches data from each linked external source
        3. Returns combined enrichment data including artist information
        
        Results are cached for performance.
      operationId: enrichArtwork
      responses:
        '200':
          description: Enrichment data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtworkEnrichment'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /external/dbpedia/artwork/{resource}:
    get:
      tags: [external]
      summary: Query DBpedia for artwork information
      description: Fetch artwork metadata from DBpedia including abstract, thumbnail, artist, and museum information.
      operationId: getDbpediaArtwork
      parameters:
        - name: resource
          in: path
          required: true
          description: "DBpedia resource name or full URI (e.g., 'Mona_Lisa')"
          schema:
            type: string
      responses:
        '200':
          description: DBpedia data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /external/wikidata/artwork/{entity}:
    get:
      tags: [external]
      summary: Query Wikidata for artwork information
      description: Fetch artwork metadata from Wikidata including image, inception date, materials, and location.
      operationId: getWikidataArtwork
      parameters:
        - name: entity
          in: path
          required: true
          description: "Wikidata Q-ID or full URI (e.g., 'Q12418')"
          schema:
            type: string
      responses:
        '200':
          description: Wikidata data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    ArtworkId:
      name: id
      in: path
      required: true
      description: Unique artwork identifier (URI or ID)
      schema:
        type: string
        example: "artwork_mona_lisa"
    EventId:
      name: eventId
      in: path
      required: true
      description: Unique provenance event identifier
      schema:
        type: string
        example: "prov_001"
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    Artwork:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          description: Unique identifier (URI)
          example: "artwork_mona_lisa"
        title:
          type: string
          description: Title of the artwork
          example: "Mona Lisa"
        artist:
          type: string
          description: Artist name
          example: "Leonardo da Vinci"
        artistUri:
          type: string
          description: Artist URI in the knowledge base
          example: "artist_davinci"
        dateCreated:
          type: string
          description: Date or period of creation
          example: "1503-1519"
        medium:
          type: string
          description: Medium/technique used
          example: "Oil on poplar panel"
        dimensions:
          type: string
          description: Physical dimensions
          example: "77 cm × 53 cm"
        description:
          type: string
          description: Description of the artwork
        imageUrl:
          type: string
          format: uri
          description: URL to artwork image
        currentLocation:
          type: string
          description: Current location/museum
          example: "Louvre Museum, Paris"
        period:
          type: string
          description: Art historical period
          example: "High Renaissance"
        style:
          type: string
          description: Artistic style
          example: "Renaissance"
        provenance:
          type: array
          items:
            $ref: '#/components/schemas/ProvenanceRecord'
          description: Provenance history (when embedded)
        externalLinks:
          $ref: '#/components/schemas/ExternalLinks'
    ArtworkCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        artist:
          type: string
        dateCreated:
          type: string
        medium:
          type: string
        dimensions:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          format: uri
        period:
          type: string
        style:
          type: string
    ArtworkUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        artist:
          type: string
        dateCreated:
          type: string
        medium:
          type: string
        dimensions:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          format: uri
        period:
          type: string
        style:
          type: string
    ArtworkListResponse:
      type: object
      required: [items, total, page, limit, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Artwork'
        total:
          type: integer
          description: Total number of artworks matching the query
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        hasMore:
          type: boolean
          description: Whether there are more pages
    ExternalLinks:
      type: object
      properties:
        dbpedia:
          type: string
          format: uri
          description: DBpedia resource URI
        wikidata:
          type: string
          format: uri
          description: Wikidata entity URI
        getty:
          type: string
          format: uri
          description: Getty AAT URI

    # ----------------------------------------
    # Provenance Schemas
    # ----------------------------------------
    ProvenanceRecord:
      type: object
      required: [id, event, date]
      properties:
        id:
          type: string
          description: Unique event identifier
          example: "prov_001"
        date:
          type: string
          description: Date of the event
          example: "1519-01-01"
        event:
          type: string
          description: Type of provenance event
          enum:
            - Creation
            - Sale
            - Gift
            - Bequest
            - Theft
            - Recovery
            - Loan
            - Return
            - Acquisition
          example: "Acquisition"
        owner:
          type: string
          description: Owner name at this point
          example: "King Francis I of France"
        location:
          type: string
          description: Location during this event
          example: "Château d'Amboise"
        description:
          type: string
          description: Additional event details
        sourceUri:
          type: string
          format: uri
          description: Source document or reference URI
        order:
          type: integer
          description: Order in the provenance chain
    ProvenanceEventCreate:
      type: object
      required: [event, date]
      properties:
        event:
          type: string
          enum: [Creation, Sale, Gift, Bequest, Theft, Recovery, Loan, Return, Acquisition]
        date:
          type: string
        owner:
          type: string
        location:
          type: string
        description:
          type: string
        sourceUri:
          type: string
          format: uri
    ProvenanceEventUpdate:
      type: object
      properties:
        event:
          type: string
          enum: [Creation, Sale, Gift, Bequest, Theft, Recovery, Loan, Return, Acquisition]
        date:
          type: string
        owner:
          type: string
        location:
          type: string
        description:
          type: string
        sourceUri:
          type: string
          format: uri

    # ----------------------------------------
    # SPARQL Schemas
    # ----------------------------------------
...
    LinkedEntity:
      type: object
      description: A linked entity with URI and label
      properties:
        uri:
          type: string
          format: uri
          description: Entity URI
        label:
          type: string
          description: Human-readable label
    ExternalQueryResponse:
      type: object
      description: Generic response for external SPARQL queries
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: string
        source:
          type: string
          enum: [dbpedia, wikidata, getty]
    DBpediaInfo:
      type: object
      description: Information retrieved from DBpedia
      properties:
        source:
          type: string
          example: "dbpedia"
        uri:
          type: string
          format: uri
        abstract:
          type: string
          description: English abstract from DBpedia
        thumbnail:
          type: string
          format: uri
          description: Thumbnail image URL
        artist:
          $ref: '#/components/schemas/LinkedEntity'
        museum:
          $ref: '#/components/schemas/LinkedEntity'
        wikidata_uri:
          type: string
          format: uri
    WikidataInfo:
      type: object
      description: Information retrieved from Wikidata
      properties:
        source:
          type: string
          example: "wikidata"
        uri:
          type: string
          format: uri
        qid:
          type: string
          description: Wikidata Q-ID
          example: "Q12418"
        image:
          type: string
          format: uri
        inception:
          type: string
          description: Date of creation
        creator:
          $ref: '#/components/schemas/LinkedEntity'
        location:
          $ref: '#/components/schemas/LinkedEntity'
        material:
          $ref: '#/components/schemas/LinkedEntity'
        genre:
          $ref: '#/components/schemas/LinkedEntity'
        movement:
          $ref: '#/components/schemas/LinkedEntity'
    GettyTerm:
      type: object
      description: Getty AAT term information
      properties:
        source:
          type: string
          example: "getty"
        uri:
          type: string
          format: uri
        prefLabel:
          type: string
          description: Preferred label in English
          example: "oil paint"
        scopeNote:
          type: string
          description: Definition or scope note
        broader:
          $ref: '#/components/schemas/LinkedEntity'
    ArtworkEnrichment:
      type: object
      description: Combined enrichment data from external sources
      required: [artwork_id]
      properties:
        artwork_id:
          type: string
          description: The artwork identifier
          example: "artwork_mona_lisa"
        dbpedia:
          $ref: '#/components/schemas/DBpediaInfo'
        wikidata:
          $ref: '#/components/schemas/WikidataInfo'
        getty:
          type: array
          items:
            $ref: '#/components/schemas/GettyTerm'
        artist_dbpedia:
          $ref: '#/components/schemas/DBpediaInfo'
        artist_wikidata:
          $ref: '#/components/schemas/WikidataInfo'

    # ----------------------------------------
    # Error Schemas
    # ----------------------------------------
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
              example: "NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Artwork not found"
            details:
              type: object
              additionalProperties: true
              description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "The requested resource was not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                field: "title"
                issue: "Title is required"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"